% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_old_new}
\alias{combine_old_new}
\title{Combine old/new name for checkboxes}
\usage{
combine_old_new(data, symbol_arrow = "->", symbol_space = " ")
}
\arguments{
\item{data}{data.frame of matched names (`darwinize_names()` output)}

\item{symbol_arrow}{character string for symbol that is used in `paste0` 
to connect old and new name}

\item{symbol_space}{character string for symbol used instead of white space}
}
\value{
data.frame of darwinized user names.
}
\description{
`combine_old_new()` is a function that combines (`paste`) 
fieldname and standard
names with unicode characters to present in checkboxes.
}
