% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinzeNames.R
\name{combineOldNew}
\alias{combineOldNew}
\title{Combine old/new name for checkboxes}
\usage{
combineOldNew(data, symbolArrow = "→", symbolSpace = " ")
}
\description{
`combineOldNew()` is a function that combines (`paste`) fieldname and standard
names with unicode characters to present in checkboxes.
}
