% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{create_renaming_checkbox}
\alias{create_renaming_checkbox}
\title{Create shiny checkbox}
\usage{
create_renaming_checkbox(id, title)
}
\arguments{
\item{id}{checkbox id}

\item{title}{a character value that specifies title of the checkbox}
}
\description{
Create shiny checkbox
}
\seealso{
Other shiny: \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_definition_hover}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
