% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinzeNames.R
\name{darwinazeNames}
\alias{darwinazeNames}
\title{Darwinize names}
\usage{
darwinazeNames(dataUser, dataDWC)
}
\arguments{
\item{dataUser}{data.frame with user data}

\item{dataDWC}{data.frame with Darwin Cloud data}
}
\value{
data.frame of darwinized user names.
}
\description{
`darwinizeNames()` is a function to darwinize given names using reference 
(ie, Darwin Cloud) dataset. It returns $Old$ and $New$ ($fieldname$ and $standard$)
name (ie, name from given user dataset that had a match in reference dataset).
}
