% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinizer.R
\name{darwinize_names}
\alias{darwinize_names}
\title{Darwinize names}
\usage{
darwinize_names(data_user = NULL, data_dwc = NULL)
}
\arguments{
\item{data_user}{data.frame with user data}

\item{data_dwc}{data.frame with Darwin Cloud data that contains at least two
columns (fieldname and standard)}
}
\value{
data.frame of darwinized user names.
}
\description{
`darwinize_names()` is a function to darwinize given names using a reference
(ie, Darwin Cloud) dataset. It returns $Old$ and $New$ ($fieldname$ and
$standard$) name (ie, name from a given user dataset that had a match in
reference dataset).
}
\examples{
darwinize_names(
  data_user = bdDwC:::data_reptiles,
  data_dwc = bdDwC:::data_darwin_cloud$data
)
}
