% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinizeNames.R
\name{darwinizeNames}
\alias{darwinizeNames}
\title{Darwinize names}
\usage{
darwinizeNames(dataUser, dataDWC)
}
\arguments{
\item{dataUser}{data.frame with user data}

\item{dataDWC}{data.frame with Darwin Cloud data}
}
\value{
data.frame of darwinized user names.
}
\description{
`darwinizeNames()` is a function to darwinize given names using reference 
(ie, Darwin Cloud) dataset. It returns $Old$ and $New$ ($fieldname$ and $standard$)
name (ie, name from given user dataset that had a match in reference dataset).
}
\examples{
darwinizeNames(dataUser = bdDwC:::dataReptiles,
               dataDWC  = bdDwC:::dataDarwinCloud$data)

}
