% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_darwin_core_info}
\alias{data_darwin_core_info}
\title{Darwin Core Info}
\format{A data frame with 203 rows and 2 columns.}
\usage{
data_darwin_core_info
}
\description{
Darwin Core Info dataset stores standard names/terms and their definitions.
This object was created using `bdDwC:::get_darwin_core_info()` function.
}
\keyword{datasets}
