% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_reptiles}
\alias{data_reptiles}
\title{Indian Reptiles}
\format{A data frame with 50 rows and 41 columns.}
\usage{
data_reptiles
}
\description{
Dataset that contains subsample of Indian reptile observations
from the iNaturalist platform.
}
\keyword{datasets}
