% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDarwinCloud}
\alias{dataDarwinCloud}
\title{Retrieve Darwin Cloud Data}
\description{
Darwin Cloud data used to Darwinize
}
\details{
Data was retrieved using `bdDwC:::getCloudData()` function and contains
Darwin Cloud data and download date.
}
