% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataReptiles}
\alias{dataReptiles}
\title{Indian Reptiles}
\format{A data frame with 50 rows and 41 columns.}
\usage{
dataReptiles
}
\description{
Dataset that contains subsample of Indian reptile observations 
from the iNaturalist platform.
}
\keyword{datasets}
