% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{download_cloud_data}
\alias{download_cloud_data}
\title{Download Darwin Cloud Data}
\usage{

  download_cloud_data(path_remote = "https://raw.githubusercontent.com/kurator-org/",
  path_github = "kurator-validation/master/packages/kurator_dwca/",
  path_file = "data/vocabularies/darwin_cloud.txt",
  column_field = "fieldname", column_stand = "standard")
}
\arguments{
\item{path_remote}{a character string that specifies path to remote
repository (Kurator in Github)}

\item{path_github}{a character string that specifies path
within given repository}

\item{path_file}{a character string that specifies name of a dictionary file}

\item{column_field}{a character string that specifies name of the column
that contains field information}

\item{column_stand}{a character string that specifies name of the column
that contains standard information}
}
\value{
data.frame of Darwin Cloud data.
}
\description{
`download_cloud_data()` is a function used to download Darwin Cloud
data. First it downloads data, then it subsets informative columns and
renames them so they could be used in `shiny` app.
}
\examples{
download_cloud_data()
}
\seealso{
Other dictionary functions: \code{\link{clean_dwc}},
  \code{\link{get_darwin_core_info}}
}
\concept{dictionary functions}
