% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCloudData.R
\name{downloadCloudData}
\alias{downloadCloudData}
\title{Download Darwin Cloud Data}
\usage{

  downloadCloudData(pathRemote = "https://raw.githubusercontent.com/kurator-org",
  pathGithub = "/kurator-validation/master/packages/kurator_dwca/data/vocabularies/",
  pathFile = "darwin_cloud.txt", columnField = "fieldname",
  columnStand = "standard")
}
\arguments{
\item{pathRemote}{Path to remote repository (Kurator in Github).}

\item{pathGithub}{Path within given repository.}

\item{pathFile}{Name of a file.}

\item{columnField}{Name of the column that contains field information.}

\item{columnStand}{Name of the column that contains standard information.}
}
\value{
data.frame of Darwin Cloud data.
}
\description{
`downloadCloudData()` is a function used to download Darwin Cloud 
data. First it downloads data, then it subsets informative columns and 
renames them so they could be used in `shiny` app.
}
\examples{
downloadCloudData()

}
