% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarwinCloudData.R
\name{getCloudData}
\alias{getCloudData}
\title{Retrieve Darwin Cloud Data}
\usage{

  getCloudData(pathCloud = "https://raw.githubusercontent.com/kurator-org/kurator-validation/master/packages/kurator_dwca/data/vocabularies/darwin_cloud.txt")
}
\arguments{
\item{pathCloud}{Path to Darwin Cloud data.}
}
\value{
NULL.
}
\description{
`getCloudData()` is a function (not exported) to download and save
Darwin Cloud data. This data is used to rename user data with function
`darwinazeNames()`. Additionally this function saves date of the download so
that this information could be shown in a `shiny` app. Therefore we create 
a list (Darwind Cloud data and time stamp (`Sys.Date()`)).
}
\examples{
\dontrun{
    getCloudData()
}

}
