% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDarwinCoreInfo.R
\name{getDarwinCoreInfo}
\alias{getDarwinCoreInfo}
\title{Retrieve Information about Darwin Core Terms}
\usage{
getDarwinCoreInfo(pathDarwinCloud = "http://tdwg.github.io/dwc/terms/")
}
\arguments{
\item{pathDarwinCloud}{Path to Darwin Cloud data.}
}
\value{
data.frame that contains name and definition for each available
Darwin Cloud term.
}
\description{
`getDarwinCoreInfo()` is a function (not exported) to download
Darwin Core Terms. This information is displayed when using manually renaming.
}
\examples{
\dontrun{
    getDarwinCoreInfo()
}

}
