% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_server_upload_databaseInput}
\alias{module_server_upload_databaseInput}
\title{Input module for {module_server_upload_database}}
\usage{
module_server_upload_databaseInput(id)
}
\description{
Input module for {module_server_upload_database}
}
\seealso{
Other shiny modules: \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_localInput}},
  \code{\link{module_server_upload_local}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
