% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_server_upload_dictionaryInput}
\alias{module_server_upload_dictionaryInput}
\title{Input module for {module_server_upload_dictionary}}
\usage{
module_server_upload_dictionaryInput(id)
}
\description{
Input module for {module_server_upload_dictionary}
}
\seealso{
Other shiny modules: \code{\link{module_server_upload_databaseInput}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_dictionary}},
  \code{\link{module_server_upload_localInput}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_dictionaryUI}},
  \code{\link{module_ui_dictionary_radiobuttons_fieldOutput}},
  \code{\link{module_ui_dictionary_radiobuttons_field}},
  \code{\link{module_ui_dictionary}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
