% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_server_upload_local}
\alias{module_server_upload_local}
\title{Module to upload local users data
(calls {shiny_server_upload_local})}
\usage{
module_server_upload_local(input, output, server, rv)
}
\arguments{
\item{rv}{reactive values}
}
\description{
Module to upload local users data
(calls {shiny_server_upload_local})
}
\seealso{
Other shiny modules: \code{\link{module_server_upload_databaseInput}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_localInput}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
