% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules_ui.R
\name{module_server_upload_local_input}
\alias{module_server_upload_local_input}
\title{Input module for {module_server_upload_local}}
\usage{
module_server_upload_local_input(id_namespace = "main",
  id_input = "path_input_data",
  label_input = "CSV / DWCA ZIP file input", mime_type = c("text/csv",
  "text/comma-separated-values,text/plain", ".csv", ".zip",
  "application/zip"))
}
\arguments{
\item{id_namespace}{a character string to be namespaced}

\item{id_input}{a character that specifies input slot}

\item{label_input}{a character string to display for a file upload}

\item{mime_type}{a character vector of MIME types}
}
\description{
This module procuces UI for local file upload
}
\seealso{
Other shiny modules: \code{\link{module_server_buttons_clean}},
  \code{\link{module_server_buttons_download}},
  \code{\link{module_server_buttons_remove}},
  \code{\link{module_server_buttons_rename}},
  \code{\link{module_server_buttons_rollback}},
  \code{\link{module_server_darwinizer_input}},
  \code{\link{module_server_darwinizer}},
  \code{\link{module_server_modal_ui}},
  \code{\link{module_server_modal}},
  \code{\link{module_server_upload_database_input}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_buttons_ui}},
  \code{\link{module_ui_buttons}},
  \code{\link{module_ui_checkbox_ui}},
  \code{\link{module_ui_checkbox}},
  \code{\link{module_ui_dictionary_radiobuttons_output}},
  \code{\link{module_ui_dictionary_radiobuttons_update}},
  \code{\link{module_ui_dictionary_radiobuttons}},
  \code{\link{module_ui_dictionary_ui}},
  \code{\link{module_ui_dictionary}},
  \code{\link{module_ui_valuebox_output}},
  \code{\link{module_ui_valuebox}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
