% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_ui_buttons}
\alias{module_ui_buttons}
\title{Module to enable/disable darwinizer buttons}
\usage{
module_ui_buttons(input, output, session, rv)
}
\arguments{
\item{rv}{reactive values}
}
\description{
Module to enable/disable darwinizer buttons
}
\seealso{
Other shiny modules: \code{\link{module_server_buttons_clean}},
  \code{\link{module_server_buttons_remove}},
  \code{\link{module_server_buttons_rename}},
  \code{\link{module_server_buttons_rollback}},
  \code{\link{module_server_darwinizerInput}},
  \code{\link{module_server_darwinizer}},
  \code{\link{module_server_modalsUI}},
  \code{\link{module_server_modals}},
  \code{\link{module_server_upload_databaseInput}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_dictionaryInput}},
  \code{\link{module_server_upload_dictionary}},
  \code{\link{module_server_upload_localInput}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_buttonsUI}},
  \code{\link{module_ui_checkboxUI}},
  \code{\link{module_ui_checkbox}},
  \code{\link{module_ui_dictionaryUI}},
  \code{\link{module_ui_dictionary_radiobuttons_fieldOutput}},
  \code{\link{module_ui_dictionary_radiobuttons_field}},
  \code{\link{module_ui_dictionary}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
