% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules_ui.R
\name{module_ui_dictionary_ui}
\alias{module_ui_dictionary_ui}
\title{UI module for {module_ui_dictionary}}
\usage{
module_ui_dictionary_ui(id_namespace, id_dic_info = "dic_info",
  id_dwc = "update_darwin_cloud",
  label_dwc = "Update Darwin Cloud dictionary",
  id_file = "path_input_dictionary",
  label_file = "Choose a personal dictionary file",
  type_file = c("text/csv", ".csv",
  "text/comma-separated-values,text/plain"))
}
\arguments{
\item{id_namespace}{a character string to be namespaced}

\item{id_dic_info}{a character string that specifies dictionary info slot}

\item{id_dwc}{a character string for Darwin Cloud dictionary update button}

\item{label_dwc}{a character string to display on update button}

\item{id_file}{a character string for dictionary file slort}

\item{label_file}{a character string to display on input file slort}

\item{type_file}{a character string that specifies allowed file types for
dictionary}
}
\description{
This module procuces UI dictionary panel
}
\seealso{
Other shiny modules: \code{\link{module_server_buttons_clean}},
  \code{\link{module_server_buttons_download}},
  \code{\link{module_server_buttons_remove}},
  \code{\link{module_server_buttons_rename}},
  \code{\link{module_server_buttons_rollback}},
  \code{\link{module_server_darwinizer_input}},
  \code{\link{module_server_darwinizer}},
  \code{\link{module_server_modal_ui}},
  \code{\link{module_server_modal}},
  \code{\link{module_server_upload_database_input}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_local_input}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_buttons_ui}},
  \code{\link{module_ui_buttons}},
  \code{\link{module_ui_checkbox_ui}},
  \code{\link{module_ui_checkbox}},
  \code{\link{module_ui_dictionary_radiobuttons_output}},
  \code{\link{module_ui_dictionary_radiobuttons_update}},
  \code{\link{module_ui_dictionary_radiobuttons}},
  \code{\link{module_ui_dictionary}},
  \code{\link{module_ui_valuebox_output}},
  \code{\link{module_ui_valuebox}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
