% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinizer.R
\name{rename_user_data}
\alias{rename_user_data}
\title{Rename Dataset According to Darwinized Names}
\usage{
rename_user_data(data_user = NULL, data_renamed = NULL)
}
\arguments{
\item{data_user}{a data.frame to be renamed}

\item{data_renamed}{a data.frame that was created using
`bdDwC::darwinize_names()` function using same `data_user`}
}
\value{
data.frame of submitted user data, but renamed according given
renames data.frame
}
\description{
`rename_user_data()` is a function used to rename given dataset
(usually given user data) according previously Darwinized Names.
}
\examples{
result <- darwinize_names(
  data_user = bdDwC:::data_reptiles,
  data_dwc = bdDwC:::data_darwin_cloud$data
)
rename_user_data(bdDwC:::data_reptiles, result)

}
\seealso{
Other darwinizer functions: 
\code{\link{darwinize_names}()},
\code{\link{link_old_new}()}
}
\concept{darwinizer functions}
