% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameUserData.R
\name{renameUserData}
\alias{renameUserData}
\title{Rename Dataset According Darwinized Names}
\usage{
renameUserData(dataUser, dataRenamed)
}
\arguments{
\item{dataUser}{data.frame of be renamed.}

\item{dataRenamed}{data.frame that was created using `bdDwC:::darwinazeNames()`
function using same `dataUser`.}
}
\value{
data.frame of submitted user data, but renamed according given 
renames data.frame.
}
\description{
`renameUserData()` is a function (not exported) to rename given dataset
(usually given user data) according previously Darwinized Names.
}
