% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_server_download_renamed}
\alias{shiny_server_download_renamed}
\title{Download renamed data}
\usage{
shiny_server_download_renamed(data_user, data_rename,
  filename = format(Sys.time(), "darwinizedData_\%Y_\%b_\%d_\%X.csv"))
}
\arguments{
\item{data_user}{a data.frame to be renamed (passed to rename_user_data)}

\item{data_rename}{a data.frame to rename by  (passed to rename_user_data)}

\item{filename}{a caracter string to use as a file name}
}
\description{
Download renamed data
}
\seealso{
Other shiny: \code{\link{shiny_server_modal_citation}},
  \code{\link{shiny_server_modal_cloud}},
  \code{\link{shiny_server_modal_dictionary}},
  \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
