% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_server_modal_custom}
\alias{shiny_server_modal_custom}
\title{Show shiny custom modal}
\usage{
shiny_server_modal_custom(title = "Wanted Title!",
  body = "Wanted Body", size = "m", easyClose = TRUE)
}
\arguments{
\item{title}{an object string passed as `shiny::modalDialog` title}

\item{body}{an object shown withing `shiny::modalDialog`}

\item{size}{a value that specifies size of the dialog box}

\item{easyClose}{logical value to specify in box is easy to close}
}
\description{
Show shiny modal with specified title, subtitle and other details
}
\seealso{
Other shiny: \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
