% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_server_modal_welcome}
\alias{shiny_server_modal_welcome}
\title{Show shiny welcoming modal}
\usage{
shiny_server_modal_welcome(title = "Welcome to Darwinizer!",
  subtitle = "Darwinize Your Data", package = "bdDwC")
}
\arguments{
\item{title}{a character string passed as `shiny::modalDialog` title}

\item{subtitle}{a character string shown withing `shiny::modalDialog`}
}
\description{
Show shiny modal with specified title, subtitle and other details
}
\seealso{
Other shiny: \code{\link{shiny_server_download_renamed}},
  \code{\link{shiny_server_modal_citation}},
  \code{\link{shiny_server_modal_cloud}},
  \code{\link{shiny_server_modal_dictionary}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
