% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_server_reactivevalues}
\alias{shiny_server_reactivevalues}
\title{Create reactive values used within app}
\usage{
shiny_server_reactivevalues()
}
\value{
An object that stores reactive values
(created with {shiny::reactiveValues})
}
\description{
Create reactive values used within app
}
\seealso{
Other shiny: \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
