% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_server_tab_darwinizer}
\alias{shiny_server_tab_darwinizer}
\title{Update activity (enable/disable) of darwinizer tab}
\usage{
shiny_server_tab_darwinizer(data_user)
}
\arguments{
\item{data_user}{a data.frame of submitted records to be darwnized}
}
\description{
Update activity (enable/disable) of darwinizer tab
}
\seealso{
Other shiny: \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
