% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_server_upload_local}
\alias{shiny_server_upload_local}
\title{Upload local users data}
\usage{
shiny_server_upload_local(path_input = NULL)
}
\arguments{
\item{path_input}{path to local file from `shiny::fileInput`}
}
\description{
Upload local users data
}
\seealso{
Other shiny: \code{\link{shiny_server_modal_citation}},
  \code{\link{shiny_server_modal_cloud}},
  \code{\link{shiny_server_modal_dictionary}},
  \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
