% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_ui_darwin_core_definition}
\alias{shiny_ui_darwin_core_definition}
\title{Extract darwin core definitions and return as ui element}
\usage{
shiny_ui_darwin_core_definition()
}
\arguments{
\item{input}{a list with arguments for quering data from the databse}
}
\description{
Extract darwin core definitions and return as ui element
}
\seealso{
Other shiny: \code{\link{shiny_server_modal_custom}},
  \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
