% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_ui_definition_hover}
\alias{shiny_ui_definition_hover}
\title{Extract darwin core definitions and return as hovering ui element}
\usage{
shiny_ui_definition_hover(names)
}
\arguments{
\item{names}{a vector of standard names}
}
\description{
Extract darwin core definitions and return as hovering ui element
}
\seealso{
Other shiny: \code{\link{create_renaming_checkbox}},
  \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
