% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_ui_dictionary}
\alias{shiny_ui_dictionary}
\title{Create information about dictionaries}
\usage{
shiny_ui_dictionary(path_dictionary = NULL, date_dictionary, ns)
}
\arguments{
\item{path_dictionary}{a character string with path to users dictionary}

\item{date_dictionary}{a value with date of darwin cloud dictionary date}

\item{ns}{namespace created with {shiny::NS}}
}
\description{
Create information about dictionaries
}
\seealso{
Other shiny: \code{\link{shiny_server_modal_custom}},
  \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
