% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_ui_dictionary_radiobuttons}
\alias{shiny_ui_dictionary_radiobuttons}
\title{Create radio buttons for user dictionary}
\usage{
shiny_ui_dictionary_radiobuttons(names_dic = NULL, inputid, label,
  spanid, selected)
}
\arguments{
\item{names_dic}{a vector that contains column names of user dictionary}

\item{inputid}{value passed as `inputId` to `shiny::radioButtons`}

\item{label}{value passed as `label` to `shiny::radioButtons`}

\item{spanid}{value used to change span id in shiny::radioButtons}

\item{selected}{selected name in radio buttons}
}
\description{
Create radio buttons for user dictionary
}
\seealso{
Other shiny: \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
