% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_cloud}
\alias{test_cloud}
\title{Test If Provided Path to Cloud Data is Valid}
\usage{
test_cloud(string = NULL)
}
\arguments{
\item{string}{a character string with path}
}
\value{
error or nothing
}
\description{
Test If Provided Path to Cloud Data is Valid
}
\seealso{
Other test functions: 
\code{\link{test_columns_cloud}()},
\code{\link{test_data_dwc}()},
\code{\link{test_data_renamed}()},
\code{\link{test_data_user}()}
}
\concept{test functions}
\keyword{internal}
