% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_columns_cloud}
\alias{test_columns_cloud}
\title{Test If Provided Columns of Cloud Data are Valid}
\usage{
test_columns_cloud(strings = NULL)
}
\arguments{
\item{string}{a vector that containts two character strings}
}
\value{
error or nothing
}
\description{
Test If Provided Columns of Cloud Data are Valid
}
\seealso{
Other test functions: 
\code{\link{test_cloud}()},
\code{\link{test_data_dwc}()},
\code{\link{test_data_renamed}()},
\code{\link{test_data_user}()}
}
\concept{test functions}
\keyword{internal}
