% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_data_dwc}
\alias{test_data_dwc}
\title{Test If Provided Dictionary Data is Valid}
\usage{
test_data_dwc(data)
}
\arguments{
\item{data}{a data.frame with dictionary data}
}
\value{
error or nothing
}
\description{
Test If Provided Dictionary Data is Valid
}
\seealso{
Other test functions: \code{\link{test_data_renamed}},
  \code{\link{test_data_user}}
}
\concept{test functions}
\keyword{internal}
