% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_data_renamed}
\alias{test_data_renamed}
\title{Test If Provided Renaming Data is Valid}
\usage{
test_data_renamed(data)
}
\arguments{
\item{data}{a data.frame with renaming data}
}
\value{
error or nothing
}
\description{
Test If Provided Renaming Data is Valid
}
\seealso{
Other test functions: \code{\link{test_data_dwc}},
  \code{\link{test_data_user}}
}
\concept{test functions}
\keyword{internal}
