% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_data_user}
\alias{test_data_user}
\title{Test If Provided User Data is Valid}
\usage{
test_data_user(data = NULL)
}
\arguments{
\item{data}{a data.frame with user data}
}
\value{
error or nothing
}
\description{
Test If Provided User Data is Valid
}
\seealso{
Other test functions: \code{\link{test_cloud}},
  \code{\link{test_columns_cloud}},
  \code{\link{test_data_dwc}},
  \code{\link{test_data_renamed}}
}
\concept{test functions}
\keyword{internal}
