% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{clean_dwc}
\alias{clean_dwc}
\title{Clean Dictionary Data}
\usage{
clean_dwc(data, column_field = "fieldname", column_stand = "standard")
}
\arguments{
\item{data}{a data.frame with dictionary data}

\item{column_field}{a character string that specifies name of the column
that contains field information}

\item{column_stand}{a character string that specifies name of the column
that contains standard information}
}
\value{
a data.frame of cleaned dictionary data
}
\description{
Clean dictionary from unnecessary or empty fields
}
\seealso{
Other dictionary functions: \code{\link{download_cloud_data}},
  \code{\link{get_darwin_core_info}}
}
\concept{dictionary functions}
\keyword{internal}
