% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDarwinCloud}
\alias{dataDarwinCloud}
\title{Darwin Cloud Data}
\format{A list of length two. First entry contains a data frame with 
422 rows and 2 columns. Second entry contains a `Date` class object.}
\usage{
dataDarwinCloud
}
\description{
Darwin Cloud data used to Darwinize. Darwin Cloud data was retrieved using 
`bdDwC::downloadCloudData()` function. This dataset contains Darwin Cloud data
and download date.
}
\keyword{datasets}
