% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarwinCloudData.R
\name{downloadCloudData}
\alias{downloadCloudData}
\title{Download Darwin Cloud Data}
\usage{
downloadCloudData(pathCloud = NULL, columnField = "fieldname",
  columnStandard = "standard")
}
\arguments{
\item{pathCloud}{Path to Darwin Cloud data.}

\item{columnField}{Name of the column that contains field information.}

\item{columnStandard}{Name of the column that contains standard information.}
}
\value{
data.frame of Darwin Cloud data.
}
\description{
`downloadCloudData()` is a function (not exported) to download Darwin Cloud 
data. First it downloads data, then it subsets informative columns and 
renames them so they could be used in `shiny` app.
}
