% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{get_darwin_core_info}
\alias{get_darwin_core_info}
\title{Retrieve Information about Darwin Core Terms}
\usage{

  get_darwin_core_info(path_darwin_cloud = "http://tdwg.github.io/dwc/terms/")
}
\arguments{
\item{path_darwin_cloud}{Path to Darwin Cloud data.}
}
\value{
data.frame that contains name and definition for each available
Darwin Cloud term.
}
\description{
`get_darwin_core_info()` is a function (not exported) to download
Darwin Core Terms. This information is displayed when using manually
renaming.
}
\seealso{
Other dictionary functions: \code{\link{clean_dwc}},
  \code{\link{download_cloud_data}}
}
\concept{dictionary functions}
\keyword{internal}
