% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_server_upload_database}
\alias{module_server_upload_database}
\title{Module to query data from the remote database
(calls {shiny_server_upload_database})}
\usage{
module_server_upload_database(input, output, server, rv)
}
\arguments{
\item{rv}{reactive values}
}
\description{
Module to query data from the remote database
(calls {shiny_server_upload_database})
}
\seealso{
Other shiny modules: \code{\link{module_server_modalsUI}},
  \code{\link{module_server_modals}},
  \code{\link{module_server_upload_databaseInput}},
  \code{\link{module_server_upload_dictionaryInput}},
  \code{\link{module_server_upload_dictionary}},
  \code{\link{module_server_upload_localInput}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_dictionaryUI}},
  \code{\link{module_ui_dictionary_radiobuttons_fieldOutput}},
  \code{\link{module_ui_dictionary_radiobuttons_field}},
  \code{\link{module_ui_dictionary}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
