% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules_ui.R
\name{module_server_upload_database_input}
\alias{module_server_upload_database_input}
\title{Input module for {module_server_upload_database}}
\usage{
module_server_upload_database_input(id_namespace = "main",
  id_text = "scientific_name", label_text = "Scientific Name:",
  value_text = "Puma concolor", id_num = "record_size",
  label_num = "Record Size:", value_num = 500,
  id_coord = "has_coords", label_coord = "Records Filter:",
  value_id_coord = 1:3, value_label_coord = c("With Coordinates",
  "Without Coordinates", "No Filter"), id_database = "query_db",
  label_database = "Online Database:", value_id_database = c("gbif",
  "idigbio", "ecoengine", "vertnet", "bison", "inat", "ala", "obis",
  "antweb"),
  value_label_database = c("GBIF (Global Biodiversity Information Facility)",
  "iDigBio (Integrated Digitized Biocollections)",
  "EcoEngine (Berkeley Ecoinformatics Engine)",
  "Vertnet (Vertebrate Network)",
  "BISON (Biodiversity Information Serving Our Nation)", "iNaturalist",
  "ALA (Atlas of Living Australia)",
  "OBIS (Ocean Biogeographic Information System)", "AntWeb"),
  id_query = "query_database", label_query = "query_database")
}
\arguments{
\item{id_namespace}{a character string to be namespaced}

\item{id_text}{a character that specifies text slot}

\item{label_text}{a character to display for a text slot}

\item{value_text}{a character that species scientific name}

\item{id_num}{a character that specifies numeric slot}

\item{label_num}{a character to display for a numeric slot}

\item{value_num}{a numeric value for number of species to download}

\item{id_coord}{a character that specifies coordinate slot}

\item{label_coord}{a character to display for a coordinate slot}

\item{value_id_coord}{numeric values as index for {value_label_coord}}

\item{value_label_coord}{a character string (vector) to display as
coordinate selections}

\item{id_database}{a character string that specifies database slot}

\item{value_id_database}{a character vector as index for
{value_label_database}}

\item{value_label_database}{a character vector to display as database names}

\item{id_query}{a character string that specifies action button to query}

\item{label_query}{a character string to display on query database button}

\item{label_databasea}{a character string to display for a database slot}
}
\description{
Input module for {module_server_upload_database}
}
\seealso{
Other shiny modules: \code{\link{module_server_buttons_clean}},
  \code{\link{module_server_buttons_download}},
  \code{\link{module_server_buttons_remove}},
  \code{\link{module_server_buttons_rename}},
  \code{\link{module_server_buttons_rollback}},
  \code{\link{module_server_darwinizer_input}},
  \code{\link{module_server_darwinizer}},
  \code{\link{module_server_modal_ui}},
  \code{\link{module_server_modal}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_local_input}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_buttons_ui}},
  \code{\link{module_ui_buttons}},
  \code{\link{module_ui_checkbox_ui}},
  \code{\link{module_ui_checkbox}},
  \code{\link{module_ui_dictionary_radiobuttons_output}},
  \code{\link{module_ui_dictionary_radiobuttons_update}},
  \code{\link{module_ui_dictionary_radiobuttons}},
  \code{\link{module_ui_dictionary_ui}},
  \code{\link{module_ui_dictionary}},
  \code{\link{module_ui_valuebox_output}},
  \code{\link{module_ui_valuebox}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
