% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{module_server_upload_local_input}
\alias{module_server_upload_local_input}
\title{Input module for {module_server_upload_local}}
\usage{
module_server_upload_local_input(id, upload_local_file = c("text/csv",
  "text/comma-separated-values,text/plain", ".csv", ".zip",
  "application/zip"))
}
\description{
Input module for {module_server_upload_local}
}
\seealso{
Other shiny modules: \code{\link{module_server_buttons_clean}},
  \code{\link{module_server_buttons_download}},
  \code{\link{module_server_buttons_remove}},
  \code{\link{module_server_buttons_rename}},
  \code{\link{module_server_buttons_rollback}},
  \code{\link{module_server_darwinizer_input}},
  \code{\link{module_server_darwinizer}},
  \code{\link{module_server_modals_ui}},
  \code{\link{module_server_modals}},
  \code{\link{module_server_upload_database_input}},
  \code{\link{module_server_upload_database}},
  \code{\link{module_server_upload_local}},
  \code{\link{module_ui_buttons_ui}},
  \code{\link{module_ui_buttons}},
  \code{\link{module_ui_checkbox_ui}},
  \code{\link{module_ui_checkbox}},
  \code{\link{module_ui_dictionary_radiobuttons_field_output}},
  \code{\link{module_ui_dictionary_radiobuttons_field}},
  \code{\link{module_ui_dictionary_ui}},
  \code{\link{module_ui_dictionary}},
  \code{\link{module_ui_valuebox_output}},
  \code{\link{module_ui_valuebox}}
}
\concept{shiny modules}
\keyword{internal}
\keyword{modules}
\keyword{shiny}
