% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rename_user_data}
\alias{rename_user_data}
\title{Rename Dataset According Darwinized Names}
\usage{
rename_user_data(data_user, data_renamed)
}
\arguments{
\item{data_user}{data.frame of be renamed.}

\item{data_renamed}{data.frame that was created using 
`bdDwC::darwinize_names()` function using same `data_user`.}
}
\value{
data.frame of submitted user data, but renamed according given 
renames data.frame.
}
\description{
`rename_user_data()` is a function used to rename given dataset
(usually given user data) according previously Darwinized Names.
}
\examples{
result <- darwinize_names(data_user = bdDwC:::data_reptiles, 
                          data_dwc  = bdDwC:::data_darwin_cloud$data)
rename_user_data(bdDwC:::data_reptiles, result)

}
