% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_dwc}
\alias{run_dwc}
\title{Launch bdDwC Shiny Application}
\usage{
run_dwc(path_shiny = "shiny")
}
\value{
`shiny::runApp()` result within browser.
}
\description{
`run_dwc` is a function that starts bdverse Darwin Cloud cleaning `shiny` app.
}
\details{
#' @param path_shiny a character string that specifies path within bdDwC
package to shiny app
}
