% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_server_upload_database}
\alias{shiny_server_upload_database}
\title{Query data from the remote database}
\usage{
shiny_server_upload_database(scientific_name, record_size, query_db,
  has_coords)
}
\arguments{
\item{scientific_name}{species scientific name}

\item{record_size}{numeric value (positive integer) of how many records to
get}

\item{query_db}{database to use}

\item{has_coords}{should records have coordinates}
}
\description{
Query data from the remote database
}
\seealso{
Other shiny: \code{\link{shiny_server_download_renamed}},
  \code{\link{shiny_server_modal_citation}},
  \code{\link{shiny_server_modal_cloud}},
  \code{\link{shiny_server_modal_dictionary}},
  \code{\link{shiny_server_modal_welcome}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_darwin_core_definition}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}},
  \code{\link{shiny_ui_valuebox}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
