% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{shiny_ui_valuebox}
\alias{shiny_ui_valuebox}
\title{Create shiny value boxes}
\usage{
shiny_ui_valuebox(value, subtitle, color)
}
\arguments{
\item{value}{value passed to `shinydashboard::valueBox` value argument}

\item{subtitle}{value passed to `shinydashboard::valueBox` subtitle argument}

\item{color}{value passed to `shinydashboard::valueBox` color argument}
}
\description{
Create shiny value boxes
}
\seealso{
Other shiny: \code{\link{create_renaming_checkbox}},
  \code{\link{shiny_server_modal}},
  \code{\link{shiny_server_reactivevalues}},
  \code{\link{shiny_server_tab_darwinizer}},
  \code{\link{shiny_server_upload_database}},
  \code{\link{shiny_server_upload_local}},
  \code{\link{shiny_ui_definition_hover}},
  \code{\link{shiny_ui_dictionary_radiobuttons}},
  \code{\link{shiny_ui_dictionary}}
}
\concept{shiny}
\keyword{internal}
\keyword{shiny}
