% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_data_dwc}
\alias{test_data_dwc}
\title{Test If Provided Dictionary Data is Valid}
\usage{
test_data_dwc(data = NULL, column_field = "fieldname",
  column_stand = "standard")
}
\arguments{
\item{data}{a data.frame with dictionary data}

\item{column_field}{a character string that specifies name of the column
that contains field information}

\item{column_stand}{a character string that specifies name of the column
that contains standard information}
}
\value{
error or nothing
}
\description{
Test If Provided Dictionary Data is Valid
}
\seealso{
Other test functions: \code{\link{test_cloud}},
  \code{\link{test_columns_cloud}},
  \code{\link{test_data_renamed}},
  \code{\link{test_data_user}}
}
\concept{test functions}
\keyword{internal}
